Add fatigue field and fullscreen quest creation
